Fixed open method in SettingsController.java , added average waiting time in ExportExcel.java

+++ src/main/java/com/kfet/generator/GeneratorController.java
@@ -30,7 +30,7 @@ public class GeneratorController {
             int fast= (int)fastSlider.getValue();
             int slow = (int)slowSlider.getValue();
             int cash = (int)cashSlider.getValue();
-            int frequency =3600*(int) Double.parseDouble(String.valueOf(customerFrequencySlider.getValue()));
+            double frequency =3600*Double.parseDouble(String.valueOf(customerFrequencySlider.getValue()));
 
             Customers customers = new Customers(nbCustomers,proba,fast,cash,slow,cpt,frequency);
             //System.out.println(Integer.parseInt(nbCustomersSpin.getValue().toString()));
